+++ core/src/main/java/cloud/xcan/angus/core/gm/application/cmd/tag/impl/WebTagCmdImpl.java
@@ -1,9 +1,16 @@
 package cloud.xcan.angus.core.gm.application.cmd.tag.impl;
 
+import static cloud.xcan.angus.core.gm.domain.operation.OperationResourceType.ORG_TAG;
+import static cloud.xcan.angus.core.gm.domain.operation.OperationType.CREATED;
+import static cloud.xcan.angus.core.gm.domain.operation.OperationType.DELETED;
+import static cloud.xcan.angus.core.gm.domain.operation.OperationType.UPDATED;
+
 import cloud.xcan.angus.api.commonlink.app.tag.WebTag;
+import cloud.xcan.angus.api.commonlink.tag.OrgTag;
 import cloud.xcan.angus.core.biz.Biz;
 import cloud.xcan.angus.core.biz.BizTemplate;
 import cloud.xcan.angus.core.biz.cmd.CommCmd;
+import cloud.xcan.angus.core.gm.application.cmd.operation.OperationLogCmd;
 import cloud.xcan.angus.core.gm.application.cmd.tag.WebTagCmd;
 import cloud.xcan.angus.core.gm.application.query.tag.WebTagQuery;
 import cloud.xcan.angus.core.gm.domain.tag.WebTagRepo;
@@ -29,6 +36,9 @@ public class WebTagCmdImpl extends CommCmd<WebTag, Long> implements WebTagCmd {
   @Resource
   private WebTagTargetRepo webTagTargetRepo;
 
+  @Resource
+  private OperationLogCmd operationLogCmd;
+
   @Transactional(rollbackFor = Exception.class)
   @Override
   public List<IdKey<Long, Object>> add(List<WebTag> tags) {
@@ -42,24 +52,27 @@ public class WebTagCmdImpl extends CommCmd<WebTag, Long> implements WebTagCmd {
 
       @Override
       protected List<IdKey<Long, Object>> process() {
-        return batchInsert(tags, "name");
+        List<IdKey<Long, Object>> idKeys = batchInsert(tags, "name");
+        operationLogCmd.addAll(ORG_TAG, tags, CREATED);
+        return idKeys;
       }
     }.execute();
   }
 
   @Transactional(rollbackFor = Exception.class)
   @Override
-  public void update(List<WebTag> webTags) {
+  public void update(List<WebTag> tags) {
     new BizTemplate<Void>() {
       @Override
       protected void checkParams() {
         // Check the tag names existed
-        webTagQuery.checkUpdateTagNameExist(webTags);
+        webTagQuery.checkUpdateTagNameExist(tags);
       }
 
       @Override
       protected Void process() {
-        batchUpdateOrNotFound(webTags);
+        List<WebTag> tagsDb =  batchUpdateOrNotFound(tags);
+        operationLogCmd.addAll(ORG_TAG, tagsDb, UPDATED);
         return null;
       }
     }.execute();
@@ -71,11 +84,13 @@ public class WebTagCmdImpl extends CommCmd<WebTag, Long> implements WebTagCmd {
     new BizTemplate<Void>() {
       @Override
       protected Void process() {
-        // Delete relation
-        webTagTargetRepo.deleteByTagIdIn(ids);
+        List<WebTag> tagsDb = webTagQuery.checkAndFind(ids);
+        if (!tagsDb.isEmpty()) {
+          webTagTargetRepo.deleteByTagIdIn(ids);
+          webTagRepo.deleteByIdIn(ids);
 
-        // Delete tags
-        webTagRepo.deleteByIdIn(ids);
+          operationLogCmd.addAll(ORG_TAG, tagsDb, DELETED);
+        }
         return null;
       }
     }.execute();
